2.0.5
First release :
PhpMyNewsLetter v2.0.5

+++ include/view_log.php
@@ -19,12 +19,9 @@ if(tok_val($token)){
 } else {
     die("3");
 }
-
 extract($_GET,EXTR_OVERWRITE);
-echo '<!DOCTYPE HTML>
-<html lang="fr">
-<head></head><body>';
-echo '<style type="text/css">body,td,th {font-size: 12px;font-family: Arial, Helvetica, sans-serif;}</style>';
+echo '<style type="text/css">body,td,th{font-size:12px;font-family:Arial,Helvetica,sans-serif;}</style>';
+echo '<div class="modal-header"><button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button></div><div class="modal-body">';
 echo '<code>';
 $temp_color='#FFFFFF';
 switch($t){
@@ -34,7 +31,7 @@ switch($t){
             $fp=fopen($log,'r');
 	        while(!feof($fp)){
 		        $ligne=fgets($fp,4096); // 4096 à modifier après confirmation MAX RECSIZE
-		        echo '<div id="item_lu" style="background-color:'.$temp_color.'">';
+		        echo '<div id="item_lu" style="background-color:'.$temp_color.';color:#000;">';
 		        echo str_replace(' ','&nbsp;',trim($ligne)).'<br>';
 		        if($temp_color=='#ECE9D8'){$temp_color='#FFFFFF';}else{$temp_color='#ECE9D8';}
 		        echo '</div>'."\n";
@@ -47,7 +44,20 @@ switch($t){
             $fp=fopen($log,'r');
 	        while(!feof($fp)){
 		        $ligne=fgets($fp,4096); // 4096 à modifier après confirmation MAX RECSIZE
-		        echo '<div id="item_lu" style="background-color:'.$temp_color.'">';
+		        echo '<div id="item_lu" style="background-color:'.$temp_color.';color:#000;">';
+		        echo str_replace(' ','&nbsp;',trim($ligne)).'<br>';
+		        if($temp_color=='#ECE9D8'){$temp_color='#FFFFFF';}else{$temp_color='#ECE9D8';}
+		        echo '</div>'."\n";
+	        }
+        }
+    break;
+    case 'u':
+        $log="../logs/$u.log";
+        if(is_file($log)){
+            $fp=fopen($log,'r');
+	        while(!feof($fp)){
+		        $ligne=fgets($fp,4096); // 4096 à modifier après confirmation MAX RECSIZE
+		        echo '<div id="item_lu" style="background-color:'.$temp_color.';color:#000;">';
 		        echo str_replace(' ','&nbsp;',trim($ligne)).'<br>';
 		        if($temp_color=='#ECE9D8'){$temp_color='#FFFFFF';}else{$temp_color='#ECE9D8';}
 		        echo '</div>'."\n";
@@ -56,7 +66,7 @@ switch($t){
     break;
 }
 echo '</code>';
-echo '</body></html>';
+echo '</div><div class="modal-footer"><button type="button" class="btn btn-default" data-dismiss="modal">Close</button></div>';
